Refactor: Modularize preference, mapPreferences and install
