+++ src/constants.js
@@ -1,0 +1,2 @@
+export const DEFAULT_STORAGE_PREFIX = 'vp';
+export const DEFAULT_REACTIVE_PROPERTIES_PREFIX = `${DEFAULT_STORAGE_PREFIX}:tracked`;
