+++ src/map-preferences.js
@@ -1,0 +1,17 @@
+import preference from './preference';
+
+function normalizeMap(map) {
+  return Array.isArray(map)
+    ? map.map(name => ({ name, options: {} }))
+    : Object.keys(map).map(name => ({ name, options: map[name] }));
+}
+
+export default function mapPreferences(preferences) {
+  const res = {};
+
+  normalizeMap(preferences).forEach(({ name, options }) => {
+    res[name] = preference(name, options);
+  });
+
+  return res;
+}
